POC to add subresource integrity
A small proof of concept to include subresource integrity. This alone is
not very tricky, but we include / load a few wasm libraries which are
loaded dynamically and would fall outside of the integrity hash.
However, a neat solution is to use the integrity option in fetch(),
which is generally used to pull in the wasm blobs anyways.

This does require a small rework of the way the wasm is initialized, but
it shouldn't be too difficult.

Ultimately this should result in the index.html loading all resources
with (cascading) integrity hashes, so that it can be signed.