Add data space invitations
Enables inviting new users by email. The new user can already be
registered or require a new account. In either case, the data space is
only notified about the new user once they accept the invite. Next, any
member can confirm the new member, so that the metadata key is shared as
well.

To confirm a new member, a hello message is sent so that the
fingerprint is prompted and can be confirmed out of band.